rename cfg_toml to data
