+++ src/data/lock/mod.rs
@@ -8,7 +8,7 @@ static LOCKFILE_HEADER: &str = r#"# Lockfile v1
 # This file is automatically generated by Wapm.
 # It is not intended for manual editing. The schema of this file may change."#;
 
-use crate::cfg_toml::manifest::MANIFEST_FILE_NAME;
+use crate::data::manifest::MANIFEST_FILE_NAME;
 use std::path::Path;
 
 pub fn is_lockfile_out_of_date<P: AsRef<Path>>(directory: P) -> Result<bool, failure::Error> {
