+++ src/dataflow/installed_manifest_packages.rs
@@ -1,4 +1,4 @@
-use crate::cfg_toml::manifest::Manifest;
+use crate::data::manifest::Manifest;
 use crate::dataflow::resolved_manifest_packages::ResolvedManifestPackages;
 use crate::dataflow::WapmPackageKey;
 use crate::util::{
